PAE-1415: Extract cell-value helpers into cell-extractor module
Splits the cell-value extraction helpers (extractCellValue, isCellEmpty
and the predicate functions they delegate to) into a sibling module so
the parser file drops below SonarCloud's per-file NCLOC threshold. The
public surface is unchanged - extractCellValue is still re-exported
from exceljs-parser.js for the existing test imports.
